[FEATURE] Add option to filter dumps

+++ src/Dumper.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace CoStack\MysqlLoader;
 
 use ArrayObject;
-use CoStack\MysqlLoader\Hacks\MysqlOrderedCreateTableSql;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\DriverManager;
 use ZipArchive;
@@ -15,7 +14,13 @@ use function fclose;
 use function file_exists;
 use function fopen;
 use function fwrite;
+use function hash_file;
+use function implode;
+use function rename;
 use function sprintf;
+use function str_starts_with;
+use function strlen;
+use function substr;
 use function unlink;
 
 class Dumper
@@ -50,6 +55,9 @@ class Dumper
         ArrayObject $files,
     ): void {
         $preamble = $dumpConfiguration->folder . '_preamble.sql';
+        if (file_exists($preamble)) {
+            rename($preamble, $preamble . '.bak');
+        }
         $handle = fopen($preamble, 'wb');
 
         $drops = $creates = $truncates = [];
@@ -91,6 +99,13 @@ class Dumper
         }
 
         fclose($handle);
+        if (
+            file_exists($preamble . '.bak')
+            && hash_file('sha1', $preamble) === hash_file('sha1', $preamble . '.bak')
+        ) {
+            unlink($preamble);
+            rename($preamble . '.bak', $preamble);
+        }
         $files[] = $preamble;
     }
 
@@ -103,16 +118,30 @@ class Dumper
         foreach ($dbInfo->nonEmptyNonExcludedTableNames as $tableName) {
             $fileName = $dumpConfiguration->folder . $tableName . '.csv';
             if (file_exists($fileName)) {
-                unlink($fileName);
+                rename($fileName, $fileName . '.bak');
             }
-            $connection->executeStatement(
-                sprintf(
-                    'SELECT * INTO OUTFILE \'%s\' FROM %s',
-                    $fileName,
-                    $tableName,
-                ),
-            );
+            $queries = [];
+            foreach ($dumpConfiguration->filterQuery as $query) {
+                if (str_starts_with($query, $tableName . ':')) {
+                    $queries[] = substr($query, strlen($tableName . ':'));
+                }
+            }
+            $where = implode(' AND ', $queries);
+            if (!empty($where)) {
+                $connection->executeStatement(
+                    sprintf('SELECT * INTO OUTFILE \'%s\' FROM %s WHERE %s', $fileName, $tableName, $where),
+                );
+            } else {
+                $connection->executeStatement(sprintf('SELECT * INTO OUTFILE \'%s\' FROM %s', $fileName, $tableName));
+            }
             $files[] = $fileName;
+            if (
+                file_exists($fileName . '.bak')
+                && hash_file('sha1', $fileName) === hash_file('sha1', $fileName . '.bak')
+            ) {
+                unlink($fileName);
+                rename($fileName . '.bak', $fileName);
+            }
         }
     }
 }
